Refactor: move Qt repo properties to QtRepoProperty
Move default_desktop_arch_dir & dir_for_version to QtRepoProperty
